Added the EntityArchetypeChunk class.
EntityArchetypeChunks are specialized collections which group entities along with their components. These EntityArchetypeChunks are typically 16 KB in size, which helps with cache locality. For performance purposes, a single byte array is used to to contain entity and component data. As such, only blittable types may be stored within an EntityArchetypeChunk. Although this restriction will prevent the storage of managed classes and structs, it does offer some performance increases through safe usage of the Unsafe class.