Added the EntityRegistry class.
EntityRegistry is a class whose purpose is to make the handling of entities and their components much easier to work with. The EntityRegistry class contains methods which creates and destroys entities with a set of component types. In addition, the EntityRegistry contains methods which allows components to be added and removed from entities at runtime, causing them to moved into a different chunk, changing their archetype in the process. In order to ensure the integrity of its internal data in a multithreaded environment, the EntityRegistry class makes use of locks at strategic locations across various methods.