Remove volatile reads/writes from EntityQuery.cs
Given that the EntityTableLookup class performs a volatile read whenever its Count property is called, it seems redundant to have the EntityQuery class perform a volatile read as well. I reordered the conditions to ensure that a volatile read from the lookup is performed first before the query's lookup index is accessed.