Add IndexOf method to EntityArchetype class
The new IndexOf method offers a quick way to find the position of a component type in the ComponentTypes property. It will first query the ComponentBitmask property to see if the archetype contains a specified component type. If the archetype does contain the component type, it will determine the range to search the component type based on its category. Afterwards, it will perform a binary search within that range and return the index of the component type.