Relax array access restrictions in EntityTable.cs
The EntityTable class will now publicly expose its entity and component arrays to callers. This is done to simplify access to the underlying data. However, great care must be taken to avoid writing data beyond the bounds of the entity table. Furthermore, the entity arrays must not be modified if the tables they originate from are managed by an entity registry.