Rename Monophyll to Logos
Monophyll was a working title I made to refer to a family of FOSS libraries I intend to create for game development using C#. The Mono- prefix derived inspiration from MonoGame, a FOSS game framework that I played around with for quite a bit while I worked on the ECS library. The -phyll suffix, meaning "of leaves," actually came from desire to create a 2D game. Although I still hold onto that desire, I became hooked on designing underlying systems that can be used to build games from the ground up. However, I was not content with making one giant library that housed all the tools developers need to make games, so I decided to separate them into submodules, hence the dot notation used in Monophyll.Entities.

The new title, Logos, is a Greek term that refers to the idea of conducting rational discourse through reasoning. I feel that it is fitting name, as I intend to develop my libraries using sound design choices. Any library under the Logos title should strive to be as modular and extensible as possible, which is why they shall be released as FOSS. This should allow developers to freely utilize these libraries in their own code, and build upon them when necessary. Logos is not just a cool sounding title for the sake of sounding smart, it is a promise stating that any tools created under this title will be developed using reasonable design choices.