Simplify API of classes with value type semantics
This commit aims to simplify the API of the ComponentType, Entity, EntityArchetype, and EntityFilter classes. Consequently, this commit removes the Equals/Compare static methods from their respective classes in favor of the Equals/CompareTo instance methods. In addition, the ComponentType class no longer implements IEquatable, as each instance is a unique per-type singleton. The Entity struct also no longer implements IComparable, as it is not used within the library.