Add AOT compatibility
I've made changes to the ComponentType and EntityTable classes to allow the library to be AOT compatible. The ComponentType class is now an abstract class that provides a method called CreateArray, which creates arrays of an arbitrary type without the use of reflection. This allows the EntityTable class to create its internal component arrays without having to use Array.CreateInstance, which requires code to be created dynamically.