Rework EntityTable storage classes
I've replaced the EntityTableLookup and EntityTableGrouping classes with the EntityLookup and EntityGrouping classes, which can be thought as the immutable versions of the old classes. As a result, the EntityRegistry and EntityQuery classes required some hefty rewrites in order to work with these changes. However, this should allow readers to see the contents of an EntityLookup instance without leaving it open to unexpected mutations.