Rework ComponentType class
I've renamed the RuntimeType and Id properties to Type and TypeId respectively. To take full advantage of the ComponentType class being abstract, I've also made the Type and Size properties abstract, and have them return the results of the typeof and sizeof operators respectively. The Size property will report the actual size of the type represented by a given ComponentType instance, so changes were made to the EntityArchetype class to ensure that tag component types don't contribute to the size of entities.