Rename ComponentType.TypeId to Index
This change reflects a similar change that was made to the Entity struct. The idea behind renaming the TypeId property to Index was to hint at its usage as an bit index in the component bitmasks used by various classes in the library. If you have read my previous commits regarding this property, you should know that I really do not like naming members after abbreviations, so it gave me a perfect excuse to make the change, for good this time.