Implement IList interfaces in EntityGrouping class
Given that the EntityGrouping class is an immutable collection that stores its elements in sequential order, I've decided to have it implement the IList interfaces. In order to maintain immutability, any method that attempts to mutate the state of an EntityGrouping instance will throw an exception immediately. However, I've added alternative methods that mimic the behavior of these interface methods through copy-on-write semantics.