Update EntityLookup class
I've simplified the API of the EntityLookup class, and I added documentation to its public facing interface. However, I am not too happy with the underlying implementation, as it has the potential to cause performance issues if there are a large number of entries that have to be reconstructed each time a grouping is added/updated/removed. I will have to run benchmarks on it to confirm my suspicions. In the event that it becomes a problem, I am thinking of reimplementing the underlying data structure as a persistent hash array mapped trie instead of a copy-on-write array.