Rename Bitmask to Bitmap
Truth be told, the "bitmasks" returned by various classes in the library are really fancy little bitmaps that can be used to quickly query the presence of component types. One major change I've made is the moving of the BitmaskOperations.Requires/Includes/Excludes methods directly to the EntityFilter.Matches method, as that was the only place where these methods were called.