Rename EntityFilter to EntityPredicate + Update EntityQuery class
I've renamed the EntityFilter class to EntityPredicate to clarify its role in the library. In addition, I've readded caching to the EntityQuery class, which is always enabled. If uncached iteration is required, the EntityRegistry class exposes an EntityLookup via the Lookup property, whose elements can be iterated and filtered using an EntityPredicate in a thread-safe manner. Be aware that the returned EntityLookup is an immutable instance, which is not updated whenever the EntityRegistry makes a change to its copy of the instance.