Update EntityPredicate+Builder classes
I've made some minor refactors to the EntityPredicate class to make the code a little more readable. In addition, I've renamed the EntityPredicate.Matches method to Test, and the EntityPredicate.Builder.Construct method to ToPredicate. The former name change did lead to a funny change to the name of its respective unit test, which went from MatchesTest to TestTest.